requirements.txt
asgiref==3.5.0
black==21.12b0
click==8.0.3
Django==4.0.1
django-crontab==0.7.1
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
psycopg2-binary==2.9.2
